backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
psutil==5.9.6

# Gemini AI Dependencies
google-generativeai>=0.3.0

# Voice Assistant Dependencies
elevenlabs
pyaudio
python-dotenv
requests